Fixing the pin backend api

+++ backend/GofishApi/Dtos/CreateCatchingPinDTO.cs
@@ -4,8 +4,8 @@ namespace GofishApi.Dtos
 {
     public class CreateCatchingPinDTO
     {
-        public required int Latitude { get; set; }
-        public required int Longitude { get; set; }
+        public required double Latitude { get; set; }
+        public required double Longitude { get; set; }
         public required string Description { get; set; }
         public required Species SpeciesType { get; set; }
         public required int HookSize { get; set; }
