+++ backend/GofishApi/Models/IPin.cs
@@ -6,8 +6,8 @@ namespace GofishApi.Models
     public interface IPin
     {
         int Id { get; set; }
-        int Latitude { get; set; }
-        int Longitude { get; set; }
+        double Latitude { get; set; }
+        double Longitude { get; set; }
         string? Description { get; set; }
         DateTime CreatedAt { get; set; }
         PinType PinType { get; set; }
